+++ CHANGELOG.md
@@ -9,6 +9,7 @@ It follows the [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) format a
 
 | Version | Date       | Related PR |
 |---------|------------|------------|
+| [1.10.3]| 2025-10-15 | [PR(#66)](https://github.com/avenirs-esr/srv-dev/pull/66)
 | [1.10.2]| 2025-10-10 | [PR(#65)](https://github.com/avenirs-esr/srv-dev/pull/65)  |
 | [1.10.2]| 2025-10-09 | [PR(#61)](https://github.com/avenirs-esr/srv-dev/pull/61)  |
 | [1.10.1]| 2025-10-07 | [PR(#60)](https://github.com/avenirs-esr/srv-dev/pull/60)  |
@@ -40,6 +41,12 @@ It follows the [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) format a
 
 
 ---
+### [1.11.0] - 2025-10-10 - [PR(#65)](https://github.com/avenirs-esr/srv-dev/pull/65)
+- clean existing APISIX routes before routes generation.
+- Add a tag on the routes: API (avenirs-portfolio-api), BACK-OFFICE (avenirs-portfolio-back-office), SPECS for swagger end points.
+- Deletion of the generated route scripts in the APISIX clean process.
+
+
 ### [1.11.0] - 2025-10-10 - [PR(#65)](https://github.com/avenirs-esr/srv-dev/pull/65)
 - Merge openapi specifications of microservices.